requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-extensions==3.1.3
django-on-heroku==1.1.2
djangorestframework==3.12.4
drf-nested-routers==0.93.4
gunicorn==20.1.0
idna==3.3
Markdown==3.3.4
multidict==5.2.0
Pillow==8.4.0
psycopg2-binary==2.9.1
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
uuid==1.30
whitenoise==5.3.0
yarl==1.7.2
